refactor: movement store 파일명을 useMovementStore로 변경 (#81)
