+++ src/features/movement/index.ts
@@ -2,7 +2,7 @@
 // 플레이어 이동 기능 관련 컴포넌트와 로직을 export
 
 // 비즈니스 로직 레이어
-export { useMovementStore } from "./model/store";
+export { useMovementStore } from "./model/useMovementStore";
 
 // UI 컴포넌트 레이어
 export { TownEngine } from "./ui/TownEngine";
